Update vsftpd with LE certs when available - pt 1

+++ scripts/install/letsencrypt.sh
@@ -121,12 +121,29 @@ if [[ $main == yes ]]; then
   sed -i "s/ssl_certificate_key .*/ssl_certificate_key \/etc\/nginx\/ssl\/${hostname}\/key.pem;/g" /etc/nginx/sites-enabled/default
 fi
 
+# Add LE certs to ZNC, if installed.
 if [[ -f /install/.znc.lock ]]; then
     # Check for LE cert, and copy it if available.
     chkhost="$(find /etc/nginx/ssl/* -maxdepth 1 -type d | cut -f 5 -d '/')"
     if [[ -n $chkhost ]]; then
-        cat /etc/nginx/ssl/"$chkhost"/{key,fullchain}.pem > /home/znc/.znc/znc.pem
-        crontab -l > newcron.txt | sed -i  "s#cron#cron --post-hook \"cat /etc/nginx/ssl/"$chkhost"/{key,fullchain}.pem > /home/znc/.znc/znc.pem\"#g" newcron.txt | crontab newcron.txt | rm newcron.txt
+        # Make sure we only find 1 SSL cert host.
+        if [[ $("$chkhost" | wc -l) -eq 1 ]]; then
+            cat /etc/nginx/ssl/"$chkhost"/{key,fullchain}.pem > /home/znc/.znc/znc.pem
+            crontab -l > newcron.txt | sed -i  "s#cron#cron --post-hook \"cat /etc/nginx/ssl/"$chkhost"/{key,fullchain}.pem > /home/znc/.znc/znc.pem\"#g" newcron.txt | crontab newcron.txt | rm newcron.txt
+        fi
+    fi
+fi
+
+# Add LE certs to VSFTPD, if installed.
+if [[ -f /install/.vsftpd.lock ]]; then
+    # Check for LE cert, and copy it if available.
+    chkhost="$(find /etc/nginx/ssl/* -maxdepth 1 -type d | cut -f 5 -d '/')"
+    if [[ -n $chkhost ]]; then
+        # Make sure we only find 1 SSL cert host.
+        if [[ $("$chkhost" | wc -l) -eq 1 ]]; then
+            sed -i "s#rsa_cert_file=/etc/ssl/certs/ssl-cert-snakeoil.pem#rsa_cert_file=/etc/nginx/ssl/${chkhost}/fullchain.pem#g" /etc/vsftpd.conf
+            sed -i "s#rsa_private_key_file=/etc/ssl/private/ssl-cert-snakeoil.key#rsa_private_key_file=/etc/nginx/ssl/${chkhost}/key.pem#g" /etc/vsftpd.conf
+        fi
     fi
 fi
